feat: Stub out frontend API calls, reorganized models, renamed routes
