+++ assets/src/models/detour.ts
@@ -1,64 +1,30 @@
 import { LatLngLiteral } from "leaflet"
 import { ShapePoint, Stop } from "../schedule"
 import { CreateDetourMachineInput } from "./createDetourMachine"
-import { array, Infer, nullable, number, string, type } from "superstruct"
+import { any, Infer, number, string, type } from "superstruct"
 
-export const SimpleDetour = type({
-  route: string(),
-  direction: string(),
-  name: string(),
-  intersection: string(),
-  updatedAt: number(),
+export const DetourWithState = type({
+  author: string(),
+  state: any(),
+  updatedAt: number()
 })
 
-export type SimpleDetour = Infer<typeof SimpleDetour>
+export type DetourWithState = Infer<typeof DetourWithState>
 
-export const SimpleDetourFromApi = type({
-  route: string(),
-  direction: string(),
-  name: string(),
-  intersection: string(),
-  updated_at: number(),
+export const DetourWithStateFromApi = type({
+  author: string(),
+  state: any(),
+  updated_at: number()
 })
 
-export type SimpleDetourFromApi = Infer<typeof SimpleDetourFromApi>
+export type DetourWithStateFromApi = Infer<typeof DetourWithStateFromApi>
 
-const simpleDetourFromData = (
-  detourData: SimpleDetourFromApi
-): SimpleDetour => ({
-  route: detourData.route,
-  direction: detourData.direction,
-  name: detourData.name,
-  intersection: detourData.intersection,
-  updatedAt: detourData.updated_at,
-})
-
-export const GroupedSimpleDetours = type({
-  active: nullable(array(SimpleDetour)),
-  draft: nullable(array(SimpleDetour)),
-  past: nullable(array(SimpleDetour)),
-})
-
-export type GroupedSimpleDetours = Infer<typeof GroupedSimpleDetours>
-
-export const GroupedDetoursFromApi = type({
-  active: nullable(array(SimpleDetourFromApi)),
-  draft: nullable(array(SimpleDetourFromApi)),
-  past: nullable(array(SimpleDetourFromApi)),
-})
-
-export type GroupedDetoursFromApi = Infer<typeof GroupedDetoursFromApi>
-
-export const groupedDetoursFromData = (
-  groupedDetours: GroupedDetoursFromApi
-): GroupedSimpleDetours => ({
-  active:
-    groupedDetours.active?.map((detour) => simpleDetourFromData(detour)) ||
-    null,
-  draft:
-    groupedDetours.draft?.map((detour) => simpleDetourFromData(detour)) || null,
-  past:
-    groupedDetours.past?.map((detour) => simpleDetourFromData(detour)) || null,
+export const detourStateFromData = (
+  detourWithState: DetourWithStateFromApi
+): DetourWithState => ({
+  author: detourWithState.author,
+  state: detourWithState.state,
+  updatedAt: detourWithState.updated_at
 })
 
 export interface DetourShape {
