+++ lib/skate_web/controllers/detours_controller.ex
@@ -24,8 +24,8 @@ defmodule SkateWeb.DetoursController do
     json(conn, %{data: returned_uuid})
   end
 
-  @spec get_detours(Plug.Conn.t(), map()) :: Plug.Conn.t()
-  def get_detours(conn, _params) do
+  @spec detours(Plug.Conn.t(), map()) :: Plug.Conn.t()
+  def detours(conn, _params) do
     %{id: user_id} = AuthManager.Plug.current_resource(conn)
 
     detours = Detours.grouped_detours(user_id)
