+++ lib/skate_web/router.ex
@@ -159,9 +159,10 @@ defmodule SkateWeb.Router do
     get "/location_search/place/:id", LocationSearchController, :get
     get "/location_search/search", LocationSearchController, :search
     get "/location_search/suggest", LocationSearchController, :suggest
+    get "/detours", DetoursController, :detours
+    get "/detours/:detour_id", DetoursController, :detour    
     post "/detours/directions/", DetourRouteController, :directions
     put "/detours/update_snapshot", DetoursController, :update_snapshot
-    get "/detours/get_detours", DetoursController, :get_detours
     post "/detours/unfinished_detour", DetoursController, :unfinished_detour
     post "/detours/finished_detour", DetoursController, :finished_detour
   end
